# Optimize R.path and dependents