--- a/source/internal/_nth.js
+++ b/source/internal/_nth.js
@@ -1,7 +1,6 @@
 import _isString from './_isString.js';
 
-function nth(offset, list) {
+export default function _nth(offset, list) {
   var idx = offset < 0 ? list.length + offset : offset;
   return _isString(list) ? list.charAt(idx) : list[idx];
 }
-export default nth;