--- a/source/internal/_path.js
+++ b/source/internal/_path.js
@@ -1,7 +1,7 @@
 import _isInteger from './_isInteger.js';
 import _nth from './_nth.js';
 
-function path(pathAr, obj) {
+export default function _path(pathAr, obj) {
   var val = obj;
   for (var i = 0; i < pathAr.length; i += 1) {
     if (val == null) {
@@ -17,5 +17,3 @@ function path(pathAr, obj) {
   }
   return val;
 }
-
-export default path;