--- a/source/last.js
+++ b/source/last.js
@@ -1,4 +1,5 @@
-import nth from './nth.js';
+import _curry1 from './internal/_curry1.js';
+import _nth from './internal/_nth.js';
 
 
 /**
@@ -21,5 +22,7 @@ import nth from './nth.js';
  *      R.last('abc'); //=> 'c'
  *      R.last(''); //=> ''
  */
-var last = nth(-1);
+var last = _curry1(function(list) {
+  return _nth(-1, list);
+});
 export default last;