--- a/source/nth.js
+++ b/source/nth.js
@@ -1,6 +1,5 @@
 import _curry2 from './internal/_curry2.js';
-import _isString from './internal/_isString.js';
-
+import _nth from './internal/_nth.js';
 
 /**
  * Returns the nth element of the given list or string. If n is negative the
@@ -28,8 +27,5 @@ import _isString from './internal/_isString.js';
  * @symb R.nth(0, [a, b, c]) = a
  * @symb R.nth(1, [a, b, c]) = b
  */
-var nth = _curry2(function nth(offset, list) {
-  var idx = offset < 0 ? list.length + offset : offset;
-  return _isString(list) ? list.charAt(idx) : list[idx];
-});
+var nth = _curry2(_nth);
 export default nth;