--- a/source/path.js
+++ b/source/path.js
@@ -1,5 +1,5 @@
 import _curry2 from './internal/_curry2.js';
-import paths from './paths.js';
+import _path from './internal/_path.js';
 
 /**
  * Retrieves the value at a given path. The nodes of the path can be arbitrary strings or non-negative integers.
@@ -26,7 +26,5 @@ import paths from './paths.js';
  *      R.path([-2], {'-2': 'a'}); //=> undefined
  */
 
-var path = _curry2(function path(pathAr, obj) {
-  return paths([pathAr], obj)[0];
-});
+var path = _curry2(_path);
 export default path;