--- a/source/pathEq.js
+++ b/source/pathEq.js
@@ -1,6 +1,6 @@
 import _curry3 from './internal/_curry3.js';
+import _path from './internal/_path.js';
 import equals from './equals.js';
-import path from './path.js';
 
 
 /**
@@ -14,7 +14,7 @@ import path from './path.js';
  * @typedefn Idx = String | Int | Symbol
  * @sig a -> [Idx] -> {a} -> Boolean
  * @param {*} val The value to compare the nested property with
- * @param {Array} path The path of the nested property to use
+ * @param {Array} pathAr The path of the nested property to use
  * @param {Object} obj The object to check the nested property in
  * @return {Boolean} `true` if the value equals the nested object property,
  *         `false` otherwise.
@@ -28,7 +28,7 @@ import path from './path.js';
  *      const isFamous = R.pathEq(90210, ['address', 'zipCode']);
  *      R.filter(isFamous, users); //=> [ user1 ]
  */
-var pathEq = _curry3(function pathEq(val, _path, obj) {
-  return equals(path(_path, obj), val);
+var pathEq = _curry3(function pathEq(val, pathAr, obj) {
+  return equals(_path(pathAr, obj), val);
 });
 export default pathEq;