My attempt at solving functions like path running slow due to using currying when not necessary and depending on functions that do more work than required.

## Results
```
path old x 2,060,753 ops/sec ±3.83% (22 runs sampled)
path new x 4,981,880 ops/sec ±0.92% (24 runs sampled)
```

## Optimization steps
- Own code for path function to avoid curry call and extra array when using paths
- moved nth functionality to internal to have access to an non curried version
- moved path functionality to internal to have access to an non curried version
- replacing calls of nth and path with non curried versions will also improve performance of dependent functions

## Curry vs no curry
In the future all functions should probably call an internal non curried version of the functionality. Not creating a new function on the fly and then calling it was a major part of these performance optimizations. Being able to use these non curried versions from the library would be amazing, although this would probably go against ramda's "everything is curried" policy.

```
path curried  x 4,981,880 ops/sec ±0.92% (24 runs sampled)
path internal x 8,866,690 ops/sec ±4.40% (23 runs sampled)
```

see [here](https://github.com/Berndy/ramda/blob/benchmark-path/benchmarkPath.md) for my benchmark experiment