@kedashoe 2023-10-31
> Any ideas what could cause this? I am both confused a switch from ternary to if would cause such a big hit in performance and why this would end up gaining performance when called in path...
> 
> see [here](https://github.com/Berndy/ramda/blob/benchmark-path/benchmarkNth.md)

I tried pulling that down and running the benchmarks but something on the branch is breaking the build for me. In any event, if a micro-optimization makes sense and the benchmarks look better, we can ship it. If either of those two conditions aren't met, we can hold off for now.