@Berndy 2023-10-31
> I tried pulling that down and running the benchmarks but something on the branch is breaking the build for me. In any event, if a micro-optimization makes sense and the benchmarks look better, we can ship it. If either of those two conditions aren't met, we can hold off for now.

My bad, I mostly hacked those benchmarks together, which broke tests (and apparently the build). Exporting anything but default apparently breaks everything, so I moved all alternative versions of code to their own files, so it should work now.

I added a whole bunch of extra variations to narrow down the cause, and I see those performance gains ONLY when it's this exact code and ONLY when it's called via a function. 

I've switched babel node to see if it's runtime based, and now I get mostly similar results for all variations. So my latest theory is now that this is the result of some bun magic that maybe optimizes very specific function patterns...?

I guess in this case it doesn't really make sense to add this to the PR.